gpa_sumcheck no more
